[MCRED-794][rw61x][zigbee] Retrieve IEEE 64 bits address from OTP memory
The IEEE 64 bits address can be programmed to OTP memory on MAC side.
This memory can be read by sending a platform request to the MAC.
This commits implements a call to the MAC API to get the IEEE address,
if it exists.
We still keep the generation of the address on host side if the address
is not programmed in OTP memory.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>