[MCSE-30][WorkQ] The OSA thread definition must be done by the user
ThreadX doesn't support dynamic stack allocation for a thread, so it
becomes complicated for the workqueue service to allocate the stack for
each thread, as we would need to detect whether FreeRTOS or ThreadX is
used. Instead, modify the API to move the thread definition to the user
side.

The user shall use `FWK_WORKQ_THREAD_DEFINE()` macro to define a thread
structure. This is a wrapper of `OSA_THREAD_DEFINE()` macro which
initially assigns a dummy thread handler. This is forced by the OSA
API. The thread handler will be overridden in `WORKQ_Start()` with
the workqueue thread handler.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>