[MCSE-674][wireless_mcu][lowpower] Fix possible array overrun
The variable `spcLowPowermode`, which is used as an array index, is
compared to the size of the `spc_lp_name_2_pd_state` array in bytes,
which in this case is evaluated to 36 bytes.
The initial intention was to compare against the array size in size of
its members, so 36 / 4 = 9.
Fixed it by using the `ARRAY_SIZE` macro.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>