[MCSE-364][Sensors] Register Temperature callbacks to enable NBU temperature handling inside ISR
Register temperature callbacks on platform (for request event and measurement ready)
this allows to trigger the temperature request after NBU request by the wireless_mcu
platform using Sensors API, and refresh the temperature when the temperature
measurement is ready.

[MCSE-364][Sensors] Protect calls to SENSORS_RefreshTemperatureValue from different tasks

Use the recursivity of the mutexes to identify the caller to SENSORS_RefreshTemperatureValue.
Lock the mutex a first time in the trigger function, then lock it again in the refresh function.
The second lock will be successful only if it's the current mutex holder.
This ensures only the mutex holder has access to the ADC.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>