[MCSE-364][Sensors] Fix SENSORS_TriggerTemperatureMeasurement API
If several trigger calls are made simultaneously (several request
from NBU to retrieve the temperature) The sensors mutex will be
locked on each call even if a measurement is ongoing. In this case
the mutex will never be free since the corresponding unlock is done
once when the measurement is ready.
Moving the Sensor mutex lock to when the trigger can be performed.