[MCSE-735][Sensors] Add support for periodic temperature measurement
The paltform can register a periodic interval on the module
to set the max interval to refresh the temperature

The purpose of this feature is to address NBU periodic temperature measurement requests.
- If interval==0 => trig oneshot measurement and stop ongoing periodic ones.
- If interval!=0 => trig a measurement and setup the new interval

Intervals are received in Ms.
The timer counter used by the module gets restarted each time the temperature is refreshed.
On k4 platforms The deviation from which the new temperature is sent to NBU is controlled by PLATFORM_TEMP_SENT_NBU_THRESHOLD macro.