[MCSE-460][wireless_mcu] Fix PLATFORM_SetXtal32MhzTrim not updating the XTAL trim when not saving to HWParam
Issue:
When calling PLATFORM_SetXtal32MhzTrim(..., FALSE), the trim value
passed in parameter is ignored, and the one from HWParam is used
instead. This behavior doesn't correspond to the function description
so it needs a fix.

Fix:
Trivial fix consisting of always referring to the trim value passed in
parameter. If `saveToHwParams` is set to TRUE, the HWParam value will
be the same than the trim value passed in parameter.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>