[MCSE-460][wireless_mcu] Disable XTAL32M temperature compensation on the host core
Changing the XTAL32M CDAC on the host core causes issues with the radio.
If the CDAC is changed during radio activity, it can disrupt the RF
activity.
For now, we disable this by putting this feature under
`gPlatformUseXtal32MTempComp` flag.
This flag can be enabled for testing only. but not for production.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>