[MCSE-936][mcxw23] Implement HCI interface using PLATFORM API
Implement the PLATFORM HCI interface as a preliminary requirement for
Zephyr enablement.
Introduce a new API PLATFORM_SendHciMessageAlt() as an alternative
API, allowing to pass the packet type as a parameter.
This will allow for an optimized HCI transport without any copy or
buffer allocation. This is useful on platforms on which the controller
runs on the same core like MCXW23.
Implement this API as not supported on platforms using an NBU.

Keep the previous logic if gUseHciTransportDownward_d is not enabled,
for HCIOT reference design applications. This will avoid having to
rework those apps which are already too mature.

This change will apply only to BLE OOBE applications.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>