[MCSE-1106][kw47_mcxw72] Record the conn_device_memory header as part of the framework
To support preprocessing on linker scripts for both building from the repo
and from a package, we have to make sure the include headers are located
at the same place between the repo and the package.
If the header is recorded in a project segment, the header will be moved
to the project folder, so the include path used in the preprocessing
command will be wrong.
To workaround that, we can record the header in a component instead of
a project segment. Here, we do it as part of the framework platform
component.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>